Adding RadioGroup & RadioOption Functionality
The RadioGroup component, the parent of the RadioOption JSX elements,
set the proper props on its children, which props will be used as props
to the <input> element.

Thus, RadioGroup using the React.Children.map() to iterate its children and
the React.cloneElement in order to acquire a copy of its children and pass
additional props.

Finally, the onChange attribute of the <input> element is a little tricky
because we need to use the onChange prop which practically is the setSelected
setState function.